+++ CHANGELOG.md
@@ -10,7 +10,9 @@ nav_order: 1
 <!-- [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
 - Add organization application users support
+- Configure "insufficient broker" error retries timeout 
 
 ## [4.12.1] - 2024-01-05